requirements.txt
praw
python-dotenv
requests
